+++ snd-data.c
@@ -802,11 +802,11 @@ void display_info(snd_info *sp)
 	  strftime(timestr, TIME_STR_SIZE, STRFTIME_FORMAT, localtime(&(sp->write_date)));
 #endif
 	  mus_snprintf(buffer, INFO_BUFFER_SIZE, 
-		       "srate: %d\nchans: %d\nlength: %.3f (%d %s)\ntype: %s\nformat: %s\nwritten: %s%s%s%s\n",
+		       "srate: %d\nchans: %d\nlength: %.3f (%ld %s)\ntype: %s\nformat: %s\nwritten: %s%s%s%s\n",
 		       hdr->srate,
 		       hdr->chans,
 		       (Float)(hdr->samples) / (Float)(hdr->chans * hdr->srate),
-		       (hdr->samples) / (hdr->chans),
+		       (long)((hdr->samples) / (hdr->chans)),
 		       (hdr->chans == 1) ? "samples" : "frames",
 		       mus_header_type_name(hdr->type),
 		       mus_data_format_name(hdr->format),
